+++ src/components/daos/DaoAgentSelector.tsx
@@ -1,13 +1,13 @@
-import { useState, useEffect, memo } from "react";
+"use client";
+
+import { useState, useEffect } from "react";
 import { Bot, Copy, Check, Send } from "lucide-react";
-import dynamic from "next/dynamic";
 import { useAgents } from "@/hooks/use-agents";
 import { useWalletStore, type WalletBalance } from "@/store/wallet";
 import { useSessionStore } from "@/store/session";
 import { useToast } from "@/hooks/use-toast";
 import { supabase } from "@/utils/supabase/client";
 import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
 import {
   Sheet,
   SheetContent,
@@ -15,112 +15,10 @@ import {
   SheetTitle,
   SheetDescription,
 } from "@/components/ui/sheet";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-  DialogDescription,
-} from "@/components/ui/dialog";
-import Image from "next/image";
-import type { Agent, Wallet, DAO, Token } from "@/types/supabase";
-
-const TokenTransfer = dynamic(
-  () =>
-    import("@/components/auth/token-transfer").then((mod) => mod.TokenTransfer),
-  {
-    ssr: false,
-  }
-);
-
-interface TransferModalProps {
-  isOpen: boolean;
-  onOpenChange: (open: boolean) => void;
-  agent: {
-    id: string;
-    name: string;
-    tokenSymbol: string;
-    dexPrincipal: string;
-    contractPrincipal: string;
-    walletAddress: string;
-  } | null;
-  onSuccess: () => void;
-  onError: (error: any) => void;
-}
-
-const TransferTokenModal = memo(
-  ({ isOpen, onOpenChange, agent, onSuccess, onError }: TransferModalProps) => {
-    const [amount, setAmount] = useState("");
-
-    useEffect(() => {
-      if (!isOpen) {
-        setAmount("");
-      }
-    }, [isOpen]);
-
-    if (!agent) return null;
-
-    const { contractAddress, contractName } = (() => {
-      const [address, name] = agent.dexPrincipal.split(".");
-      // Remove the "-dex" suffix from the contract name
-      const cleanedContractName = name.endsWith("-dex")
-        ? name.slice(0, -4)
-        : name;
-      return { contractAddress: address, contractName: cleanedContractName };
-    })();
-
-    const amountInMicroTokens = Number(amount) * 1000000;
-
-    const handleTransferSuccess = () => {
-      console.log({
-        amount: amountInMicroTokens,
-        contractName,
-        tokenSymbol: agent.tokenSymbol,
-        sender: contractAddress,
-        receiver: agent.walletAddress,
-      });
-      onSuccess();
-    };
-
-    return (
-      <Dialog open={isOpen} onOpenChange={onOpenChange}>
-        <DialogContent>
-          <DialogHeader>
-            <DialogTitle>Transfer Tokens to {agent.name}</DialogTitle>
-            <DialogDescription>
-              Transfer {agent.tokenSymbol} tokens
-            </DialogDescription>
-          </DialogHeader>
-          <div className="space-y-4">
-            <Input
-              type="number"
-              placeholder="Enter amount"
-              value={amount}
-              onChange={(e) => setAmount(e.target.value)}
-            />
-            <TokenTransfer
-              network={
-                process.env.NEXT_PUBLIC_STACKS_NETWORK === "mainnet"
-                  ? "mainnet"
-                  : "testnet"
-              }
-              amount={amountInMicroTokens}
-              recipient={agent.walletAddress}
-              contractAddress={contractAddress}
-              contractName={contractName}
-              token={agent.tokenSymbol}
-              buttonText="Transfer"
-              onSuccess={handleTransferSuccess}
-              onError={onError}
-            />
-          </div>
-        </DialogContent>
-      </Dialog>
-    );
-  }
-);
-
-TransferTokenModal.displayName = "TransferTokenModal";
+import { AgentAvatar } from "../reusables/AgentAvatar";
+import { AgentBalance } from "../reusables/AgentBalance";
+import { TransferTokenModal } from "./TransferTokenModal";
+import type { Wallet, DAO, Token } from "@/types/supabase";
 
 interface AgentSelectorSheetProps {
   selectedAgentId: string | null;
@@ -132,6 +30,10 @@ interface AgentSelectorSheetProps {
   tokens?: Token[];
 }
 
+/**
+ * AgentSelectorSheet Component
+ * Main component for selecting and managing agents
+ */
 export function AgentSelectorSheet({
   selectedAgentId,
   onSelect,
@@ -141,6 +43,7 @@ export function AgentSelectorSheet({
   daos,
   tokens,
 }: AgentSelectorSheetProps) {
+  // Hooks and state
   const { agents, loading: agentsLoading } = useAgents();
   const {
     balances,
@@ -150,6 +53,7 @@ export function AgentSelectorSheet({
   } = useWalletStore();
   const { userId } = useSessionStore();
   const { toast } = useToast();
+
   const [copiedAddress, setCopiedAddress] = useState<string | null>(null);
   const [error, setError] = useState<string | null>(null);
   const [transferModalOpen, setTransferModalOpen] = useState(false);
@@ -165,6 +69,7 @@ export function AgentSelectorSheet({
   // Filter out archived agents
   const activeAgents = agents.filter((agent) => !agent.is_archived);
 
+  // Fetch wallets on component mount
   useEffect(() => {
     if (userId) {
       fetchWallets(userId).catch((err) => {
@@ -174,48 +79,7 @@ export function AgentSelectorSheet({
     }
   }, [userId, fetchWallets]);
 
-  const createTasks = async (
-    agentId: string,
-    tokenSymbol: string,
-    daoMission: string
-  ) => {
-    const tasks = [
-      {
-        name: "Send Proposal",
-        prompt: `Send a new proposal to the ${tokenSymbol} based on their ${daoMission}`,
-        agent_id: agentId,
-        is_scheduled: true,
-        cron: "0 */12 * * *", // Every 12 hours
-        profile_id: userId,
-      },
-      {
-        name: "Vote on proposal",
-        prompt: `Vote on the proposal for ${tokenSymbol} which you like`,
-        agent_id: agentId,
-        is_scheduled: true,
-        cron: "30 */12 * * *", // Every 12 hours, 30 minutes after sending proposal
-        profile_id: userId,
-      },
-    ];
-
-    try {
-      const { error } = await supabase.from("tasks").insert(tasks);
-      if (error) throw error;
-
-      toast({
-        title: "Success",
-        description: "Tasks created successfully",
-      });
-    } catch (error) {
-      console.error("Error creating tasks:", error);
-      toast({
-        title: "Error",
-        description: "Failed to create tasks. Please try again.",
-        variant: "destructive",
-      });
-    }
-  };
-
+  // Helper functions
   const truncateAddress = (address: string) => {
     if (!address) return "";
     return `${address.slice(0, 5)}...${address.slice(-5)}`;
@@ -240,10 +104,6 @@ export function AgentSelectorSheet({
     }
   };
 
-  const formatBalance = (balance: string) => {
-    return (Number(balance) / 1_000_000).toFixed(6);
-  };
-
   const getWalletAddress = (wallet: Wallet) => {
     return process.env.NEXT_PUBLIC_STACKS_NETWORK === "mainnet"
       ? wallet.mainnet_address
@@ -255,17 +115,55 @@ export function AgentSelectorSheet({
     requiredTokenSymbol?: string
   ) => {
     if (!requiredTokenSymbol) return true;
-
-    const hasRequiredToken = Object.entries(balance.fungible_tokens).some(
-      ([tokenId, token]) => {
-        const [, tokenSymbol] = tokenId.split("::");
-        return tokenSymbol === requiredTokenSymbol && Number(token.balance) > 0;
-      }
-    );
-
-    return hasRequiredToken;
+    return Object.entries(balance.fungible_tokens).some(([tokenId, token]) => {
+      const [, tokenSymbol] = tokenId.split("::");
+      return tokenSymbol === requiredTokenSymbol && Number(token.balance) > 0;
+    });
   };
 
+  // Create tasks for the selected agent
+  const createTasks = async (
+    agentId: string,
+    tokenSymbol: string,
+    daoMission: string
+  ) => {
+    const tasks = [
+      {
+        name: "Send Proposal",
+        prompt: `Send a new proposal to the ${tokenSymbol} based on their ${daoMission}`,
+        agent_id: agentId,
+        is_scheduled: true,
+        cron: "0 */12 * * *",
+        profile_id: userId,
+      },
+      {
+        name: "Vote on proposal",
+        prompt: `Vote on the proposal for ${tokenSymbol} which you like`,
+        agent_id: agentId,
+        is_scheduled: true,
+        cron: "30 */12 * * *",
+        profile_id: userId,
+      },
+    ];
+
+    try {
+      const { error } = await supabase.from("tasks").insert(tasks);
+      if (error) throw error;
+      toast({
+        title: "Success",
+        description: "Tasks created successfully",
+      });
+    } catch (error) {
+      console.error("Error creating tasks:", error);
+      toast({
+        title: "Error",
+        description: "Failed to create tasks. Please try again.",
+        variant: "destructive",
+      });
+    }
+  };
+
+  // Event handlers
   const handleSelect = async (agentId: string | null) => {
     if (agentId && requiredTokenSymbol) {
       const dao = daos.find((d) => {
@@ -325,6 +223,7 @@ export function AgentSelectorSheet({
     });
   };
 
+  // Loading and error states
   if (error) {
     return (
       <div className="flex h-11 w-auto items-center justify-center rounded-full bg-destructive/10 text-destructive px-4">
@@ -422,56 +321,10 @@ export function AgentSelectorSheet({
                   </div>
 
                   {balance && (
-                    <div className="space-y-1">
-                      <div className="flex justify-between text-sm">
-                        <span className="text-muted-foreground">STX</span>
-                        <span className="text-foreground">
-                          {formatBalance(balance.stx.balance)}
-                        </span>
-                      </div>
-
-                      {Object.entries(balance.fungible_tokens).map(
-                        ([tokenId, token]) => {
-                          const [, tokenSymbol] = tokenId.split("::");
-                          return (
-                            <div
-                              key={tokenId}
-                              className={`flex justify-between text-sm ${
-                                requiredTokenSymbol === tokenSymbol
-                                  ? "font-bold text-foreground"
-                                  : ""
-                              }`}
-                            >
-                              <span className="text-muted-foreground">
-                                {tokenSymbol}
-                              </span>
-                              <span className="text-foreground">
-                                {formatBalance(token.balance)}
-                              </span>
-                            </div>
-                          );
-                        }
-                      )}
-
-                      {Object.entries(balance.non_fungible_tokens).map(
-                        ([tokenId, token]) => {
-                          const [, tokenSymbol] = tokenId.split("::");
-                          return (
-                            <div
-                              key={tokenId}
-                              className="flex justify-between text-sm"
-                            >
-                              <span className="text-muted-foreground">
-                                {tokenSymbol || "NFT"}
-                              </span>
-                              <span className="text-foreground">
-                                {token.count} items
-                              </span>
-                            </div>
-                          );
-                        }
-                      )}
-                    </div>
+                    <AgentBalance
+                      balance={balance}
+                      requiredTokenSymbol={requiredTokenSymbol}
+                    />
                   )}
                 </div>
               );
@@ -487,40 +340,5 @@ export function AgentSelectorSheet({
         onError={handleTransferError}
       />
     </>
-  );
-}
-
-function AgentAvatar({
-  agent,
-  className = "",
-}: {
-  agent?: Agent;
-  className?: string;
-}) {
-  if (!agent?.image_url) {
-    return (
-      <div
-        className={`flex items-center justify-center rounded-full bg-background ${className}`}
-      >
-        <Bot className="h-5 w-5 text-foreground/50" />
-      </div>
-    );
-  }
-
-  return (
-    <div className={`relative rounded-full overflow-hidden ${className}`}>
-      <Image
-        src={agent.image_url || "/placeholder.svg"}
-        alt={agent.name}
-        height={24}
-        width={24}
-        className="object-cover"
-      />
-      <div className="absolute inset-0 bg-black/20 flex items-center justify-center">
-        <span className="text-lg font-bold text-white">
-          {agent.name.charAt(0).toUpperCase()}
-        </span>
-      </div>
-    </div>
   );
 }
